+++ files/init-ejbca.sh
@@ -4,15 +4,18 @@
 #
 set -e
 
-if [ "$1" == 'ejbca' ]; then
-  # Suck in startup functions + variables
-  . /usr/local/lib/init.sh
+# Suck in startup functions + variables
+. /usr/local/lib/ejbca.sh
+. /usr/local/lib/wildfly.sh
 
-  # Start MySQL server
-  mysql_start
-
-  # Start Wildfly server
-  wildfly_start
+if [ "$1" == 'run' ]; then
+  ejbca_startup_check
+  wildfly_startup_check
+  ejbca_start
+elif [ "$1" == 'init' ]; then
+  ejbca_init
+elif [ "$1" == 'update-config' ]; then
+  ejbca_update_config
 else
   # Execute the command
   exec "$@"