+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Release History:
 ================
 
+**6.5.0-r1** _(Released 06/17/2017)_
+- Separated out MySQL server
+- Changed base image to **dataknightllc/wildfly:stable**
+- Changed initialization scripts to use secrets folder
+<br /><br />
+
 **6.5.0-r0** _(Released 06/02/2017)_
 - Initial release of the container image
 - EJBCA Community version 6.5.0.5